app/requirements.txt
Django==4.2.11
djangorestframework==3.15.0
djangorestframework-simplejwt~=5.3.1
psycopg2-binary==2.9.9
requests==2.31.0
